Move IOCServiceExtension into Disuku.Discord
