+++ Disuku.InversionOfControl/Container.cs
@@ -18,25 +18,5 @@ namespace Disuku.InversionOfControl
             .AddSingleton<IDisukuLogger, DisukuLogger>()
             .AddSingleton<ITmdbService, TmdbService>()
             .AddSingleton<IRaiderIOService, RaiderIOService>();
-
-        public static IServiceCollection AutoAddServices(this IServiceCollection services)
-        {
-            foreach (var type in Assembly.GetExecutingAssembly().GetTypes()
-                .Where(x => typeof(IServiceExtention).IsAssignableFrom(x) && !x.IsInterface))
-            {
-                services.AddSingleton(type);
-            }
-            return services;
-        }
-
-        public static async Task InitializeServicesAsync(this IServiceProvider services)
-        {
-            foreach (var type in Assembly.GetExecutingAssembly().GetTypes()
-                .Where(x => typeof(IServiceExtention).IsAssignableFrom(x) && !x.IsInterface))
-            {
-                await ((IServiceExtention)services.GetRequiredService(type)).InitializeAsync();
-            }
-        }
-
     }
 }
