+++ DisukuBot/DisukuDiscord/Extensions/IServiceExtention.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Disuku.InversionOfControl
+namespace Disuku.Discord.DisukuDiscord.Extensions
 {
     public interface IServiceExtention
     {
